Create the 'xyLOGIX.Coinbase.CDP.Tokens.Models' module
Provides objects and functionality to model the data that typically can be found in .json files distributed by the Coinbase Developer Platform (CDP) upon the generation of a CDP API key.  These files contain the name of the key and the .pem data for the user's private key.

This module is a component of the xyLOGIX Coinbase REST API Support software system.

---
Copyright © 2024 by xyLOGIX, LLC.  All rights reserved.