Create the 'xyLOGIX.Coinbase.CDP.Tokens' module
The Coinbase Developer Platform gives builders the most trusted, scaled, and reliable onramps and infrastructure available for building onchain apps.  Battle tested.

This module provides objects and functionality to seamlessly generate JSON Web Token(s) (JWTs) for use with requests to the Coinbase (Retail) and Coinbase (Advanced Trade) APIs.  The JWTs are to be generated per each API request and are to be supplied as an 'Authorization Bearer' token.
This module includes functionality to read a user-specified JSON file containing a token name and a private key, or to specify them as arguments to a method.

This module is a component of the xyLOGIX Coinbase REST API Support software system.

---
Copyright © 2024 by xyLOGIX, LLC.  All rights reserved.